Abstract ProjectTask command handler
