+++ src/core/Codend.Application/ProjectTasks/Commands/UpdateProjectTask/UpdateProjectTaskCommand.cs
@@ -1,20 +1,11 @@
 using Codend.Application.Core.Abstractions.Data;
 using Codend.Application.Core.Abstractions.Messaging.Commands;
 using Codend.Contracts.ProjectTasks;
-using Codend.Domain.Core.Errors;
 using Codend.Domain.Entities;
 using Codend.Domain.Repositories;
 using Codend.Shared.ShouldUpdate;
-using FluentResults;
 
 namespace Codend.Application.ProjectTasks.Commands.UpdateProjectTask;
-
-public interface IUpdateProjectTaskCommand<out TUpdateProjectTaskProperties>
-    where TUpdateProjectTaskProperties : UpdateProjectTaskProperties
-{
-    ProjectTaskId TaskId { get; }
-    TUpdateProjectTaskProperties UpdateTaskProperties { get; }
-}
 
 public sealed record UpdateProjectTaskCommand
 (
@@ -22,34 +13,15 @@ public sealed record UpdateProjectTaskCommand
     UpdateProjectTaskProperties UpdateTaskProperties
 ) : ICommand, IUpdateProjectTaskCommand<UpdateProjectTaskProperties>;
 
-public class UpdateProjectTaskCommandHandler<TCommand, TProjectTask, TUpdateProjectTaskProperties>
-    : ICommandHandler<TCommand>
-    where TCommand : ICommand, IUpdateProjectTaskCommand<TUpdateProjectTaskProperties>
-    where TProjectTask : ProjectTask, IProjectTaskUpdater<TProjectTask, TUpdateProjectTaskProperties>
-    where TUpdateProjectTaskProperties : UpdateProjectTaskProperties
+public class UpdateProjectTaskCommandHandler :
+    AbstractUpdateProjectTaskCommandHandler<
+        UpdateProjectTaskCommand,
+        ProjectTask,
+        UpdateProjectTaskProperties>
 {
-    private readonly IProjectTaskRepository _taskRepository;
-    private readonly IUnitOfWork _unitOfWork;
-
-    protected UpdateProjectTaskCommandHandler(IProjectTaskRepository taskRepository, IUnitOfWork unitOfWork)
+    public UpdateProjectTaskCommandHandler(IProjectTaskRepository taskRepository, IUnitOfWork unitOfWork)
+        : base(taskRepository, unitOfWork)
     {
-        _taskRepository = taskRepository;
-        _unitOfWork = unitOfWork;
-    }
-
-    public async Task<Result> Handle(TCommand request, CancellationToken cancellationToken)
-    {
-        if (await _taskRepository.GetByIdAsync(request.TaskId) is not TProjectTask task)
-        {
-            return Result.Fail(new DomainErrors.ProjectTaskErrors.ProjectTaskNotFound());
-        }
-
-        task.Update(request.UpdateTaskProperties);
-
-        _taskRepository.Update(task);
-        await _unitOfWork.SaveChangesAsync(cancellationToken);
-
-        return Result.Ok();
     }
 }
